move SMVerifier and remove dup methods from merge
